requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0
google-genai>=1.0.0
cryptography>=41.0.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
bcrypt>=4.0.0
fpdf2>=2.7.0
streamlit-autorefresh>=1.0.1
streamlit-option-menu>=0.3.12
python-docx>=1.0.0
requests>=2.28.0
python-pptx>=0.6.23
flask>=3.0.0
flask-cors>=4.0.0
pyjwt>=2.8.0
psycopg2-binary>=2.9.0
reportlab>=4.0.0
gunicorn>=22.0.0; sys_platform != 'win32'
python-telegram-bot>=20.0
streamlit-aggrid>=1.0.0
streamlit-timeline>=0.0.2
streamlit-lottie>=0.0.5
streamlit-extras>=0.4.0
